Add Widget::textBaseline() to fix several misalignments using scalable fonts
* Each widget can customize its text baseline (onGetTextBaseline()) to
  know where to draw text when it's vertically aligned to the middle
  (now more correctly "aligned to the baseline").
* Add PAINT_BASELINE to test where the baseline is on each widget.
* "Aseprite" and "Aseprite Mini" fonts now have a special
  descent/ascent configuration with a specific baseline position.

+++ src/app/fonts/fonts.h
@@ -36,7 +36,7 @@ public:
   void addFontData(const std::string& name, std::unique_ptr<FontData>&& fontData);
 
   FontData* fontDataByName(const std::string& name);
-  text::FontRef fontByName(const std::string& name, int size);
+  text::FontRef fontByName(const std::string& name, float size);
   text::FontRef fontFromInfo(const FontInfo& fontInfo);
   FontInfo infoFromFont(const text::FontRef& font);
